+++ app/controllers/conditionController.js
@@ -1,17 +1,22 @@
 const conditionDataMapper = require("../dataMappers/conditionDataMapper");
-
 const conditionController = {
-  
   // Récupère tous les états d'un manga de la base de données
   getAllConditions: async (request, response, next) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
+
       const conditions = await conditionDataMapper.findAllConditions();
       if (!conditions) {
         return next();
       }
-      response.status(200).json(
-        conditions
-      );
+      response.status(200).json(conditions);
     } catch (error) {
       console.log(error);
       return response.json({
@@ -27,21 +32,25 @@ const conditionController = {
   // Crée un nouvel état de manga dans la base de données
   createOneCondition: async (request, response) => {
     try {
-      const {
-        condition_name,
-      } = request.body;
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
+      const { condition_name } = request.body;
 
       // Vérifie la présence de tous les paramètres nécessaires dans le corps de la requête
-      if (
-        !condition_name
-      ) {
+      if (!condition_name) {
         return response.json({
-          status : 400, 
-          error: "Paramètre manquant dans le corps de la requête HTTP" 
+          status: 400,
+          error: "Paramètre manquant dans le corps de la requête HTTP"
         });
       }
       const newCondition = await conditionDataMapper.insertOneCondition({
-        condition_name,
+        condition_name
       });
 
       if (newCondition) {
@@ -58,8 +67,7 @@ const conditionController = {
         return response.json({
           status: 200,
           success: false,
-          message:
-            "Aucun état n'a été créé, peut-être que l'état existe déjà"
+          message: "Aucun état n'a été créé, peut-être que l'état existe déjà"
         });
       }
     } catch (error) {
@@ -79,7 +87,6 @@ const conditionController = {
     try {
       const { id } = request.params;
 
-  
       const condition = await conditionDataMapper.findOneConditionById(id);
       if (!condition) {
         // Aucun état trouvé, renvoyer une réponse 404 Not Found
@@ -89,9 +96,7 @@ const conditionController = {
           message: "Aucun état trouvé avec le code id spécifié"
         });
       }
-      return response.status(200).json(
-        condition
-      );
+      return response.status(200).json(condition);
     } catch (error) {
       console.log(error);
       return response.json({
@@ -104,14 +109,19 @@ const conditionController = {
     }
   },
 
-  modifyOneConditionById : async (request, response) => {
+  modifyOneConditionById: async (request, response) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
       const { id } = request.params;
 
-  
-      const {
-        condition_name
-      } = request.body;
+      const { condition_name } = request.body;
 
       if (!condition_name) {
         return response.json({
@@ -121,35 +131,32 @@ const conditionController = {
       }
       const modifiedCondition = await conditionDataMapper.updateOneCondition({
         id,
-        condition_name,
+        condition_name
       });
 
       if (modifiedCondition) {
-      // La modification s'est bien déroulée
+        // La modification s'est bien déroulée
         return response.json({
           status: 201,
           success: true,
           message: "L'état a été modifié avec succès",
-          condition: modifiedCondition,
+          condition: modifiedCondition
         });
-
       } else {
         // Aucune ligne affectée, la modification n'a pas été effectuée
         return response.json({
-          status : 200,
+          status: 200,
           success: false,
-          message: "Aucun état n'a été modifié",
+          message: "Aucun état n'a été modifié"
         });
-
       }
-
-    }catch (error) {
+    } catch (error) {
       console.log(error);
       return response.json({
-        status : 500,
+        status: 500,
         success: false,
         error: {
-          message : error.toString()
+          message: error.toString()
         }
       });
     }
@@ -158,9 +165,16 @@ const conditionController = {
   // Supprime un état par son code ISBN
   removeOneConditionById: async (request, response) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
       const { id } = request.params;
 
-  
       const condition = await conditionDataMapper.deleteOneConditionById(id);
       if (!condition) {
         // Aucun état trouvé, renvoyer une réponse 404 Not Found
