+++ app/controllers/roleController.js
@@ -4,13 +4,19 @@ const roleController = {
   // Récupère toutes les roles d'un utilisateur de la base de données
   getAllRoles: async (request, response, next) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
       const roles = await roleDataMapper.findAllRoles();
       if (!roles) {
         return next();
       }
-      response.status(200).json(
-        roles
-      );
+      response.status(200).json(roles);
     } catch (error) {
       console.log(error);
       return response.json({
@@ -26,21 +32,25 @@ const roleController = {
   // Crée un nouveau rôle dans la base de données
   createOneRole: async (request, response) => {
     try {
-      const {
-        role_name,
-      } = request.body;
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
+      const { role_name } = request.body;
 
       // Vérifie la présence de tous les paramètres nécessaires dans le corps de la requête
-      if (
-        !role_name
-      ) {
+      if (!role_name) {
         return response.json({
-          status : 400, 
-          error: "Paramètre manquant dans le corps de la requête HTTP" 
+          status: 400,
+          error: "Paramètre manquant dans le corps de la requête HTTP"
         });
       }
       const newRole = await roleDataMapper.insertOneRole({
-        role_name,
+        role_name
       });
 
       if (newRole) {
@@ -57,8 +67,7 @@ const roleController = {
         return response.json({
           status: 200,
           success: false,
-          message:
-            "Aucun rôle n'a été créé, peut-être que le rôle existe déjà"
+          message: "Aucun rôle n'a été créé, peut-être que le rôle existe déjà"
         });
       }
     } catch (error) {
@@ -78,7 +87,6 @@ const roleController = {
     try {
       const { id } = request.params;
 
-  
       const role = await roleDataMapper.findOneRoleById(id);
       if (!role) {
         // Aucun role trouvé, renvoyer une réponse 404 Not Found
@@ -88,9 +96,7 @@ const roleController = {
           message: "Aucun rôle trouvé avec le code id spécifié"
         });
       }
-      return response.status(200).json(
-        role
-      );
+      return response.status(200).json(role);
     } catch (error) {
       console.log(error);
       return response.json({
@@ -104,14 +110,20 @@ const roleController = {
   },
 
   // Modifie un rôle par son id
-  modifyOneRoleById : async (request, response) => {
+  modifyOneRoleById: async (request, response) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
       const { id } = request.params;
 
-      const {
-        role_name
-      } = request.body;
-      
+      const { role_name } = request.body;
+
       if (!role_name) {
         return response.json({
           status: 400,
@@ -124,31 +136,28 @@ const roleController = {
       });
 
       if (modifiedRole) {
-      // La modification s'est bien déroulée
+        // La modification s'est bien déroulée
         return response.json({
           status: 201,
           success: true,
-          message: 'le rôle a été modifié avec succès',
-          role: modifiedRole,
+          message: "le rôle a été modifié avec succès",
+          role: modifiedRole
         });
-
       } else {
         // Aucune ligne affectée, la modification n'a pas été effectuée
         return response.json({
-          status : 200,
+          status: 200,
           success: false,
-          message: "Aucun rôle n'a été modifié",
+          message: "Aucun rôle n'a été modifié"
         });
-
       }
-
-    }catch (error) {
+    } catch (error) {
       console.log(error);
       return response.json({
-        status : 500,
+        status: 500,
         success: false,
         error: {
-          message : error.toString()
+          message: error.toString()
         }
       });
     }
@@ -157,6 +166,14 @@ const roleController = {
   // Supprime un rôle par son code id
   removeOneRoleById: async (request, response) => {
     try {
+      // Vérifier si l'utilisateur a le rôle d'administrateur
+      if (request.user?.role_id !== 2) {
+        // Si l'utilisateur n'est pas authentifié en tant qu'administrateur, renvoyer une réponse d'erreur
+        return response.status(403).json({
+          success: false,
+          message: "Accès interdit. Vous n'êtes pas administrateur."
+        });
+      }
       const { id } = request.params;
 
       const role = await roleDataMapper.deleteOneRoleById(id);
